requirements.txt
spacy==3.7.2
pyspark==3.4.3
configparser==6.0.1
kafka-python==2.0.2
python-dotenv==1.0.1
newsapi-python==0.2.7
